+++ jest.config.js
@@ -11,4 +11,8 @@ module.exports = {
     "!**/*.d.ts",
   ],
   testPathIgnorePatterns: ["/node_modules/", "/.expo/"],
+  moduleNameMapper: {
+    "^react-native-safe-area-context$": "<rootDir>/__mocks__/react-native-safe-area-context.js",
+    "^react-native-pager-view$": "<rootDir>/__mocks__/react-native-pager-view.js",
+  },
 };